feat(agraria): implement searchable maintenance and update views for maintenance management
