--- resources/views/agraria/manutenzioni/nuova.blade.php
@@ -1,11 +1,0 @@
-@extends("layout.main")
-
-@section("title", "Manutenzione")
-
-@section("content")
-    @include("partials.header", ["title" => "Nuova Manutenzione"])
-    <nuova-manutenzione-form
-        programmate-url="{{ route("api.manutenzioni.programmate") }}"
-        mezzi-url="{{ route("api.mezzi") }}"
-    ></nuova-manutenzione-form>
-@endsection
